Separate js and css files
